♻️  refactor: 편집 서버 패치로 수정
