+++ app/(route)/admin/notice-edit/[shop_id]/[notice_id]/layout.tsx
@@ -1,15 +1,16 @@
 export const metadata = {
-    title: "공고 수정",
-  };
-  
-  export default function Layout({ children }: { children: React.ReactNode }) {
-    return (
-      <div className="mx-auto min-h-screen w-full max-w-[964px] px-3 pb-[60px] pt-[142px] md:pb-[80px] md:pt-[130px]">
-        <header className="h- flex w-full justify-between">
-          <h2 className="font-bold text-l leading-[25px]">가게 정보</h2>
-        </header>
+  title: "공고 수정",
+};
+
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="mt-[70px]">
+      <div className="ml-auto mr-auto mt-[70px] max-w-[964px] pb-20 pl-3 pr-3 pt-[60px] md:pl-8 md:pr-8">
+        <div className="mb-6 flex justify-between">
+          <h1 className="font-bold text-l">공고 수정</h1>
+        </div>
         {children}
       </div>
-    );
-  }
-  
+    </div>
+  );
+}
